pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shortlist"
version = "0.1.0"
description = "Job search chief of staff — automated daily briefings"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
browser = [
    "playwright>=1.40",
]

[project.scripts]
shortlist = "shortlist.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
